+++ web-wallet/src/lib/components/GasSettings/GasSettings.svelte
@@ -2,7 +2,7 @@
 	import { slide } from "svelte/transition";
 	import { Button } from "$lib/dusk/components";
 	import { GasControls, GasFee } from "$lib/components";
-	import { createEventDispatcher } from "svelte";
+	import { createEventDispatcher, onMount } from "svelte";
 
 	/** @type {number} */
 	export let limit;
@@ -26,6 +26,20 @@
 	let isExpanded = false;
 
 	const dispatch = createEventDispatcher();
+
+	onMount(() => {
+		let inputPrice = false;
+		let	inputLimit = false;
+		let validGasLimits = false;
+
+		inputPrice = !!(price >= priceLower && price <= limitUpper);
+
+		inputLimit = !!(limit >= limitLower && limit <= limitUpper);
+
+		validGasLimits = !!(inputPrice && inputLimit);
+
+		dispatch("checkGasLimits", validGasLimits);
+	});
 </script>
 
 <div class="gas-settings">
@@ -46,6 +60,7 @@
 		<div in:slide|global class="gas-settings">
 			<GasControls
 				on:setGasSettings={(event) => { dispatch("setGasSettings", event.detail); }}
+				on:checkGasLimits={(event) => {dispatch("checkGasLimits", event.detail); }}
 				{limit}
 				{limitLower}
 				{limitUpper}