+++ web-wallet/src/lib/components/Send/Send.svelte
@@ -2,11 +2,11 @@
 
 <script>
 	import { fade } from "svelte/transition";
-	import { createEventDispatcher, onMount } from "svelte";
+	import { createEventDispatcher, onMount, tick } from "svelte";
 	import { mdiArrowUpBoldBoxOutline, mdiWalletOutline } from "@mdi/js";
 
 	import { deductLuxFeeFrom } from "$lib/contracts";
-	import { luxToDusk } from "$lib/dusk/currency";
+	import { duskToLux, luxToDusk } from "$lib/dusk/currency";
 	import { logo } from "$lib/dusk/icons";
 	import {
 		AnchorButton,
@@ -55,10 +55,16 @@
 	/** @type {HTMLInputElement | null} */
 	let amountInput;
 
+	/** @type {boolean} */
 	let isNextButtonDisabled = false;
 
-	const checkAmountValid = () => {
-		isNextButtonDisabled = !amountInput?.checkValidity();
+	/** @type {boolean} */
+	let validGasLimits = false;
+
+	const checkAmountValid = async () => {
+		await tick();
+		isNextButtonDisabled = !(amountInput?.checkValidity() && validGasLimits
+			&& (luxFee + luxToDusk(amount) <= duskToLux(spendable)));
 	};
 
 	const dispatch = createEventDispatcher();
@@ -130,6 +136,10 @@
 					price={gasSettings.gasPrice}
 					priceLower={gasSettings.gasPriceLower}
 					on:setGasSettings
+					on:checkGasLimits={(event) => {
+						validGasLimits = event.detail;
+						checkAmountValid();
+					}}
 				/>
 			</div>
 		</WizardStep>